requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.3
CacheControl==0.13.1
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
dnspython==2.3.0
eventlet==0.33.3
firebase-admin==6.3.0
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-marshmallow==1.1.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.3.3
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
msgpack==1.0.7
multidict==6.0.4
openai==1.2.2
packaging==23.2
pdfminer.six==20221105
proto-plus==1.23.0
protobuf==4.25.2
ProWritingAidSDK @ git+https://github.com/prowriting/prowritingaid.python.git@fd6fc238778b7b5ba4bb56835d3ed73594f3bcb7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.8.0
python-socketio==5.10.0
redis==5.0.1
requests==2.31.0
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.2
